[HUDI-2081] Move schema util tests out from TestHiveSyncTool (#3166)
